cleanup: separate init functions into its own file
