+++ .gitattributes
@@ -6,6 +6,7 @@
 
 *.cpp text eol=lf
 *.cc text eol=lf
+*.h text eol=lf
 
 *.sh text eol=lf
 *.bat text eol=crlf
